removing indirection and circular reference
