+++ event-watcher/package.json
@@ -42,7 +42,7 @@
     "add": "2.0.6",
     "jest": "29.3.1",
     "prettier": "2.8.1",
-    "ts-jest": "29.0.3",
+    "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
     "typescript": "^5.2.2"
   }
