+++ event-watcher/src/watchers/BaseWatcher.ts
@@ -5,11 +5,10 @@ import { DEFAULT_RPS, NETWORK_RPS_BY_CHAIN, TIMEOUT } from '../consts';
 import { DBOptionTypes, WHTransaction, VaasByBlock, WHTransferRedeemed } from '../databases/types';
 import { getLogger, WormholeLogger } from '../utils/logger';
 import { SNSOptionTypes } from '../services/SNS/types';
-import { WatcherImplementation } from './types';
 import axios from 'axios';
 import rateLimit, { type RateLimitedAxiosInstance } from 'axios-rate-limit';
 
-abstract class BaseWatcher implements WatcherImplementation {
+abstract class BaseWatcher {
   public logger: WormholeLogger;
   maximumBatchSize: number = 100;
   sns?: SNSOptionTypes;
